slightly working

+++ src/draw/include/vector_utils.h
@@ -5,6 +5,6 @@ t_vector	add_vector(t_vector a, t_vector b);
 t_vector	substract_vector(t_vector a, t_vector b);
 t_vector	multip_vector(t_vector a, int num);
 t_vector	cross_product(t_vector a, t_vector b);
-int			dot_product(t_vector a, t_vector b);
+double		dot_product(t_vector a, t_vector b);
 
 #endif
