+++ src/draw/vector_utils.c
@@ -40,7 +40,7 @@ t_vector	cross_product(t_vector a, t_vector b)
 	return (result);
 }
 
-int	dot_product(t_vector a, t_vector b)
+double	dot_product(t_vector a, t_vector b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
